Move CascadeClient into Wrapper namespace and rename to CascadeWrapper
